# contracts/oar-openapi.yaml
openapi: 3.1.0
info:
  title: Organization Autorunner HTTP API
  version: 0.2.3
  description: |
    Canonical HTTP contract for organization-autorunner.

    This OpenAPI contract is the monorepo source of truth for generated clients,
    CLI command registry metadata, and command docs.
servers:
  - url: http://127.0.0.1:8000
...
paths:
...
  /docs:
    get:
      operationId: listDocuments
      summary: List documents and their current head metadata
      tags: [docs]
      x-oar-surface: "canonical"
      x-oar-command-id: "docs.list"
      x-oar-cli-path: "docs list"
      x-oar-why: "Discover available documents without resolving each head individually, optionally scoped to a single thread."
      x-oar-examples:
        - title: List documents
          command: oar docs list --json
        - title: Search documents by title
          command: oar docs list --q "constitution" --json
        - title: Paginated document list
          command: oar docs list --limit 50 --json
      x-oar-input-mode: "none"
      x-oar-streaming: { mode: none }
      x-oar-output-envelope: "Returns `{ documents, next_cursor? }` ordered by `updated_at` descending. Pagination is optional and backward-compatible."
      x-oar-error-codes: [invalid_request]
      x-oar-concepts: [docs, revisions]
      x-oar-stability: "beta"
      x-oar-agent-notes: "Safe and idempotent. Use `thread_id` to focus on one thread's docs and `include_tombstoned=true` when auditing superseded documents. Optional pagination with `q` for search, `limit` for page size, and `cursor` for continuation."
      parameters:
        - in: query
          name: q
          schema: { type: string }
          description: "Optional search query to filter documents by ID or title using case-insensitive matching."
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 1000 }
          description: "Maximum number of results to return. Defaults to returning all results when omitted."
        - in: query
          name: cursor
          schema: { type: string }
          description: "Opaque pagination cursor from a previous response's `next_cursor` field."
        - in: query
          name: thread_id
          schema: { type: string }
          description: "Return only documents currently linked to the given thread id."
        - in: query
          name: include_tombstoned
          schema: { type: boolean }
          description: "Include tombstoned documents in the list. Defaults to false."
      responses:
        '200':
          description: Document list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentsResponse'
    post:
      operationId: createDocument
      summary: Create document with initial immutable revision
      tags: [docs]
      x-oar-surface: "canonical"
      x-oar-command-id: "docs.create"
      x-oar-cli-path: "docs create"
      x-oar-why: "Bootstrap a first-class document identity and initial revision without manual head-pointer management."
      x-oar-examples:
        - title: Create document
          command: oar docs create --from-file doc-create.json --json
      x-oar-input-mode: "json-body"
      x-oar-streaming: { mode: none }
      x-oar-output-envelope: "Returns `{ document, revision }` where `revision` is the new head."
      x-oar-error-codes: [invalid_json, invalid_request, unknown_actor_id, conflict]
      x-oar-concepts: [docs, revisions]
      x-oar-stability: "beta"
      x-oar-agent-notes: "Replay-safe when `request_key` is reused with the same body; core can issue the canonical document id when one is omitted."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDocumentRequest'
      responses:
        '201':
          description: Document created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentWithRevisionResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
  /docs/{document_id}:
    get:
      operationId: getDocument
      summary: Get document and authoritative head revision
      tags: [docs]
      x-oar-surface: "canonical"
      x-oar-command-id: "docs.get"
      x-oar-cli-path: "docs get"
      x-oar-why: "Resolve the current authoritative document head without client-side lineage traversal."
      x-oar-examples:
        - title: Get document head
          command: oar docs get --document-id product-constitution --json
      x-oar-input-mode: "none"
      x-oar-streaming: { mode: none }
      x-oar-output-envelope: "Returns `{ document, revision }` where `revision` is the current head."
      x-oar-error-codes: [invalid_request, not_found]
      x-oar-concepts: [docs, revisions]
      x-oar-stability: "beta"
      x-oar-agent-notes: "Safe and idempotent."
      parameters:
        - $ref: '#/components/parameters/DocumentID'
      responses:
        '200':
          description: Document and head revision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentWithRevisionResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
    patch:
      operationId: updateDocument
      summary: Create a new immutable revision for an existing document
      tags: [docs]
      x-oar-surface: "canonical"
      x-oar-command-id: "docs.update"
      x-oar-cli-path: "docs update"
      x-oar-why: "Append a revision and atomically advance document head with optimistic concurrency."
      x-oar-examples:
        - title: Update document
          command: oar docs update --document-id product-constitution --from-file doc-update.json --json
      x-oar-input-mode: "json-body"
      x-oar-streaming: { mode: none }
      x-oar-output-envelope: "Returns `{ document, revision }` for the newly-created head revision."
      x-oar-error-codes: [invalid_json, invalid_request, unknown_actor_id, conflict, not_found]
      x-oar-concepts: [docs, revisions, concurrency]
      x-oar-stability: "beta"
      x-oar-agent-notes: "Set `if_base_revision` from `docs.get` to prevent lost updates."
      parameters:
        - $ref: '#/components/parameters/DocumentID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDocumentRequest'
      responses:
        '200':
          description: Document updated with new revision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentWithRevisionResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
  /docs/{document_id}/history:
    get:
      operationId: listDocumentHistory
      summary: List ordered immutable revisions for a document
      tags: [docs]
      x-oar-surface: "canonical"
      x-oar-command-id: "docs.history"
      x-oar-cli-path: "docs history"
      x-oar-why: "Traverse full document lineage in canonical revision-number order."
      x-oar-examples:
        - title: List document history
          command: oar docs history --document-id product-constitution --json
      x-oar-input-mode: "none"
      x-oar-streaming: { mode: none }
      x-oar-output-envelope: "Returns `{ document_id, revisions }` ordered by ascending `revision_number`."
      x-oar-error-codes: [invalid_request, not_found]
      x-oar-concepts: [docs, revisions, lineage]
      x-oar-stability: "beta"
      x-oar-agent-notes: "Safe and idempotent."
      parameters:
        - $ref: '#/components/parameters/DocumentID'
      responses:
        '200':
          description: Document revision history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentHistoryResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /docs/{document_id}/revisions/{revision_id}:
    get:
      operationId: getDocumentRevision
      summary: Get one immutable document revision
      tags: [docs]
      x-oar-surface: "canonical"
      x-oar-command-id: "docs.revision.get"
      x-oar-cli-path: "docs revision get"
      x-oar-why: "Read a specific historical revision payload without mutating document head."
      x-oar-examples:
        - title: Get revision
          command: oar docs revision get --document-id product-constitution --revision-id 019f... --json
      x-oar-input-mode: "none"
      x-oar-streaming: { mode: none }
      x-oar-output-envelope: "Returns `{ revision }` including metadata and revision content."
      x-oar-error-codes: [invalid_request, not_found]
      x-oar-concepts: [docs, revisions]
      x-oar-stability: "beta"
      x-oar-agent-notes: "Safe and idempotent."
      parameters:
        - $ref: '#/components/parameters/DocumentID'
        - $ref: '#/components/parameters/RevisionID'
      responses:
        '200':
          description: Document revision payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentRevisionResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
...
  /docs/{document_id}/tombstone:
    post:
      operationId: tombstoneDocument
      summary: Tombstone a document (soft-delete)
      tags: [docs]
      x-oar-surface: "canonical"
      x-oar-command-id: "docs.tombstone"
      x-oar-cli-path: "docs tombstone"
      x-oar-why: "Mark a document as inactive while preserving revision history and provenance."
      x-oar-examples:
        - title: Tombstone document
          command: oar docs tombstone --document-id product-constitution --reason "replaced by v2" --json
      x-oar-input-mode: "json-body"
      x-oar-streaming: { mode: none }
      x-oar-output-envelope: "Returns `{ document, revision }` with updated tombstone metadata."
      x-oar-error-codes: [invalid_json, invalid_request, not_found]
      x-oar-concepts: [docs, lifecycle]
      x-oar-stability: "beta"
      x-oar-agent-notes: "Idempotent; repeated tombstone calls on the same document are safe."
      parameters:
        - $ref: '#/components/parameters/DocumentID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TombstoneRequest'
      responses:
        '200':
          description: Document tombstoned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentWithRevisionResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  parameters:
...
    DocumentID:
      in: path
      name: document_id
      required: true
      schema: { type: string }
    RevisionID:
      in: path
      name: revision_id
      required: true
      schema: { type: string }
...
  responses:
    ErrorResponse:
      description: Error envelope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    AnyObject:
      type: object
      additionalProperties: true
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            recoverable: { type: boolean }
            hint: { type: string }
          required: [code, message, recoverable, hint]
      required: [error]
...
    Document:
      type: object
      description: Canonical long-lived document lineage with a mutable head revision and immutable history.
      additionalProperties: true
    DocumentRevision:
      type: object
      description: Immutable revision entry within a document lineage.
      additionalProperties: true
...
    CreateDocumentRequest:
      type: object
      properties:
        actor_id: { type: string }
        request_key: { type: string }
        document:
          $ref: '#/components/schemas/AnyObject'
        refs:
          type: array
          items: { type: string }
        content:
          oneOf:
            - type: string
            - type: object
              additionalProperties: true
        content_type:
          type: string
          enum: [text, structured, binary]
      required: [document, content, content_type]
    UpdateDocumentRequest:
      type: object
      properties:
        actor_id: { type: string }
        document:
          $ref: '#/components/schemas/AnyObject'
        if_base_revision: { type: string }
        refs:
          type: array
          items: { type: string }
        content:
          oneOf:
            - type: string
            - type: object
              additionalProperties: true
        content_type:
          type: string
          enum: [text, structured, binary]
      required: [if_base_revision, content, content_type]
    DocumentWithRevisionResponse:
      type: object
      properties:
        document:
          $ref: '#/components/schemas/Document'
        revision:
          $ref: '#/components/schemas/DocumentRevision'
      required: [document, revision]
    DocumentsResponse:
      type: object
      properties:
        documents:
          type: array
          items:
            $ref: '#/components/schemas/Document'
        next_cursor:
          type: string
          description: "Opaque pagination cursor for fetching the next page of results. Omitted when there are no more results."
      required: [documents]
    DocumentHistoryResponse:
      type: object
      properties:
        document_id: { type: string }
        revisions:
          type: array
          items:
            $ref: '#/components/schemas/DocumentRevision'
      required: [document_id, revisions]
    DocumentRevisionResponse:
      type: object
      properties:
        revision:
          $ref: '#/components/schemas/DocumentRevision'
      required: [revision]
...
    TombstoneRequest:
      type: object
      properties:
        actor_id: { type: string }
        reason: { type: string, description: "Human-readable reason for tombstoning" }
      required: [actor_id]
...